requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pip==24.3.1
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
setuptools==58.0.4
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.2
typing-extensions==4.12.2
uv==0.5.1
uvicorn==0.32.0
